% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.m2.R
\name{init.mard.2}
\alias{init.mard.2}
\title{Epidemic Model Initial Conditions for MARDHAM Models}
\usage{
init.mard.2(nwstats, prev.B = 0.275, prev.W = 0.275,
  init.prev.age.slope.B = 0.05/12, init.prev.age.slope.W = 0.05/12, ...)
}
\arguments{
\item{nwstats}{Target statistics for the network model. An object of class
\code{nwstats} output from \code{\link{calc_nwstats.mard.m2}}.}

\item{prev.B}{Initial disease prevalence among black MSM.}

\item{prev.W}{Initial disease prevalence among white MSM.}

\item{init.prev.age.slope.B}{Slope of initial prevalence by age for black MSM.}

\item{init.prev.age.slope.W}{Slope of initial prevalence by age for white MSM.}

\item{...}{Additional arguments passed to function.}
}
\value{
A list object of class \code{init.mard.m2}, which can be passed to EpiModel
function \code{\link{netsim}}.
}
\description{
Sets the initial conditions for a stochastic epidemic models
             simulated with \code{\link{netsim}}.
}

